refactor: migrate monetary fields to BigDecimal and introduce OrderStatus enum
- Replace Double/Float with BigDecimal for all money fields
- Add OrderStatus enum replacing raw string status
- Add PIX payment amount validation and confirmation endpoint
- Add delivery state machine with valid transition map
- Add transactional event publishing for WebSocket notifications

+++ backend/src/main/java/com/delivery/dto/PixResponseDTO.java
@@ -1,9 +1,11 @@
 package com.delivery.dto;
 
+import java.math.BigDecimal;
+
 public record PixResponseDTO(
     String qrCode,
     String copyPaste,
     String transactionId,
     String expiresAt,
-    Double amount
+    BigDecimal amount
 ) {}
